+++ src/Customer/Service/Password.php
@@ -10,29 +10,20 @@ declare(strict_types=1);
 namespace OxidEsales\GraphQL\Storefront\Customer\Service;
 
 use OxidEsales\GraphQL\Base\Service\Authentication;
+use OxidEsales\GraphQL\Storefront\Customer\Exception\InvalidEmail;
 use OxidEsales\GraphQL\Storefront\Customer\Exception\PasswordMismatch;
+use OxidEsales\GraphQL\Storefront\Customer\Infrastructure\Password as PasswordInfrastructure;
 use OxidEsales\GraphQL\Storefront\Customer\Service\Customer as CustomerService;
 use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\Repository;
 
 final class Password
 {
-    /** @var Repository */
-    private $repository;
-
-    /** @var CustomerService */
-    private $customerService;
-
-    /** @var Authentication */
-    private $authenticationService;
-
     public function __construct(
-        Repository $repository,
-        CustomerService $customerService,
-        Authentication $authenticationService
+        private readonly Repository $repository,
+        private readonly CustomerService $customerService,
+        private readonly Authentication $authenticationService,
+        private readonly PasswordInfrastructure $passwordInfrastructure,
     ) {
-        $this->repository = $repository;
-        $this->customerService = $customerService;
-        $this->authenticationService = $authenticationService;
     }
 
     public function change(string $old, string $new): bool
@@ -50,5 +41,25 @@ final class Password
         $customerModel->setPassword($new);
 
         return $this->repository->saveModel($customerModel);
+    }
+
+    public function sendPasswordForgotEmail(string $email): bool
+    {
+        $isSuccess = $this->passwordInfrastructure->sendPasswordForgotEmail($email);
+
+        if ($isSuccess === false) {
+            throw InvalidEmail::byString($email);
+        }
+
+        return $isSuccess;
+    }
+
+    public function resetPassword(string $updateId, string $newPassword, string $repeatPassword): bool
+    {
+        if (!$this->passwordInfrastructure->checkPassword($newPassword, $repeatPassword)) {
+            return false;
+        }
+
+        return $this->passwordInfrastructure->resetPassword($updateId, $newPassword);
     }
 }